src/translator.py
# translator.py
# Módulo para traducir texto del inglés al español.

# NOTA IMPORTANTE: Para una implementación real, necesitarías integrar una API de traducción
# como la de Google Translate (con su librería correspondiente) o la API de DeepL.
# Para este ejemplo, utilizaremos una traducción simulada.

def traducir_texto(texto):
    """
    Traduce un texto del inglés al español (simulación).

    Args:
        texto (str): El texto en inglés a traducir.

    Returns:
        str: El texto traducido al español.
    """
    # Implementación simulada de la traducción
    # ¡REEMPLAZAR ESTO CON UNA API DE TRADUCCIÓN REAL EN UN PROYECTO SERIO!
    traducciones_simuladas = {
        "This is a sample text.": "Este es un texto de ejemplo.",
        "Hello, world!": "¡Hola, mundo!",
        "The quick brown fox jumps over the lazy dog.": "El veloz zorro marrón salta sobre el perro perezoso."
    }
    return traducciones_simuladas.get(texto, f"[TRADUCCIÓN SIMULADA: {texto}]")

if __name__ == '__main__':
    texto_ingles = input("Introduce el texto en inglés que quieres traducir: ")
    texto_traducido = traducir_texto(texto_ingles)
    print(f"Texto original: {texto_ingles}")
    print(f"Texto traducido: {texto_traducido}")